requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dotenv==0.9.9
firebase-admin==6.7.0
Flask==3.1.0
gcloud==0.18.3
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jws==0.1.3
MarkupSafe==3.0.2
msgpack==1.1.0
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dotenv==1.1.0
python-jwt==2.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==79.0.0
six==1.17.0
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
